Fixed indexing issue

+++ Activity_Viewer/spine_volume.py
@@ -79,7 +79,7 @@ def get_corrected_roi_pixels(parent):
     # Arange all the artifact frames to exclude
     a_frames = parent.parameters["Artifact Frames"]
     artifact_frames = [np.linspace(x[0], x[1], x[1] - x[0] + 1) for x in a_frames]
-    artifact_frames = np.concatenate(artifact_frames)
+    artifact_frames = np.concatenate(artifact_frames).astype(int)
     all_frames = np.ones(parent.activity_trace["Spine"].shape[0])
     all_frames[artifact_frames] = 0
     good_frames = np.nonzero(all_frames == 1)[0]
@@ -131,7 +131,7 @@ def get_uncorrected_roi_pixels(parent):
     """Helper function to get the roi pixels from the average projection"""
     a_frames = parent.parameters["Artifact Frames"]
     artifact_frames = [np.linspace(x[0], x[1], x[1] - x[0] + 1) for x in a_frames]
-    artifact_frames = np.concatenate(artifact_frames)
+    artifact_frames = np.concatenate(artifact_frames).astype(int)
     all_frames = np.ones(parent.activity_trace["Spine"].shape[0])
     all_frames[artifact_frames] = 0
     good_frames = np.nonzero(all_frames == 1)[0]
